chore(ci): autofix homeboy from main

+++ src/commands/runner/doctor.rs
@@ -753,8 +753,8 @@ mod probes {
         }
         let stat = unsafe { stat.assume_init() };
         let block_size = stat.f_frsize.max(1);
-        let total_blocks = u64::from(stat.f_blocks);
-        let available_blocks = u64::from(stat.f_bavail);
+        let total_blocks = stat.f_blocks;
+        let available_blocks = stat.f_bavail;
         Some(DiskProbe {
             path: common::display_path(path),
             total_mb: total_blocks.saturating_mul(block_size) / 1024 / 1024,
